Reorganizing code, splitting up huge handlers.go file
