+++ .travis.yml
@@ -9,9 +9,8 @@ install:
   - go get github.com/ngageoint/seed-common/...
 
 script:
-   - echo "skipping tests"
    - ./build-silo.sh
-   - go test .
+   - go test ./...
 
 notifications:
   webhooks:
