+++ build-silo.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-go build -o silo main.go database.go handlers.go logger.go routes.go router.go constants.go
+go build -o silo main.go
