--- constants.go
@@ -1,3 +1,0 @@
-package main
-
-const TokenSecret = "Y4Y7)j)999>s(vDk"
