+++ run-silo.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-go run main.go database.go handlers.go logger.go routes.go router.go constants.go
+go run main.go
