+++ test/registries/registries_test.go
@@ -1,7 +1,8 @@
-package main
+package registries
 
 import (
 	"bytes"
+	"database/sql"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -12,23 +13,36 @@ import (
 	"os"
 	"testing"
 
-	"github.com/ngageoint/seed-common/objects"
+	"github.com/gorilla/mux"
 	"github.com/ngageoint/seed-common/util"
 	"github.com/ngageoint/seed-silo/models"
+	"github.com/ngageoint/seed-silo/database"
+	"github.com/ngageoint/seed-silo/route"
 )
 
 var token = ""
+var db *sql.DB
+var router *mux.Router
 
 func TestMain(m *testing.M) {
+	var err error
 	os.Remove("./silo-test.db")
-	db = InitDB("./silo-test.db")
-	router, err = NewRouter()
+	db = database.InitDB("./silo-test.db")
+	router, err = route.NewRouter()
+	if err != nil {
+		os.Remove("./silo-test.db")
+		os.Exit(-1)
+	}
 
 	util.InitPrinter(util.PrintErr)
 	log.SetFlags(0)
 	log.SetOutput(ioutil.Discard)
 
 	token, err = login("admin", "spicy-pickles17!")
+	if err != nil {
+		os.Remove("./silo-test.db")
+		os.Exit(-1)
+	}
 
 	code := m.Run()
 
@@ -45,6 +59,8 @@ func TestEmptyTable(t *testing.T) {
 	}{
 		{"/registries"},
 		{"/images"},
+		{"/jobs"},
+		{"/job-versions"},
 	}
 
 	for _, c := range cases {
@@ -71,13 +87,17 @@ func TestGetNonExistentItem(t *testing.T) {
 	}{
 		{"/registries/1/scan", 401, false, "Missing authorization token"},
 		{"/registries/1/scan", 404, true, "No registry found with that ID"},
+		{"/images/1", 404, false, "No image found with that ID"},
 		{"/images/1/manifest", 404, false, "No image found with that ID"},
+		{"/users/2", 404, false, "No user found with that ID"},
+		{"/jobs/1", 404, false, "No job found with that ID"},
+		{"/job-versions/1", 404, false, "No job version found with that ID"},
 	}
 
 	for _, c := range cases {
 		req, _ := http.NewRequest("GET", c.urlStr, nil)
 		if c.auth {
-			req.Header.Set("Authorization", "Token: " + token)
+			req.Header.Set("Authorization", "Token: "+token)
 		}
 		response := executeRequest(req)
 
@@ -97,7 +117,7 @@ func TestAddRegistry(t *testing.T) {
 	payload := []byte(`{"name":"dockerhub", "url":"https://hub.docker.com", "org":"johnptobe", "username":"", "password": ""}`)
 
 	req, _ := http.NewRequest("POST", "/registries/add", bytes.NewBuffer(payload))
-	req.Header.Set("Authorization", "Token: " + token)
+	req.Header.Set("Authorization", "Token: "+token)
 	response := executeRequest(req)
 
 	checkResponseCode(t, http.StatusCreated, response.Code)
@@ -131,13 +151,13 @@ func TestDeleteRegistry(t *testing.T) {
 
 	payload := []byte(``)
 	req, _ := http.NewRequest("DELETE", "/registries/delete/1", bytes.NewBuffer(payload))
-	req.Header.Set("Authorization", "Token: " + token)
+	req.Header.Set("Authorization", "Token: "+token)
 	response := executeRequest(req)
 
 	checkResponseCode(t, http.StatusOK, response.Code)
 
 	req, _ = http.NewRequest("GET", "/registries/1/scan", bytes.NewBuffer(payload))
-	req.Header.Set("Authorization", "Token: " + token)
+	req.Header.Set("Authorization", "Token: "+token)
 	response = executeRequest(req)
 
 	checkResponseCode(t, http.StatusNotFound, response.Code)
@@ -151,7 +171,7 @@ func TestDeleteRegistry(t *testing.T) {
 	}
 
 	req, _ = http.NewRequest("DELETE", "/registries/delete/test", bytes.NewBuffer(payload))
-	req.Header.Set("Authorization", "Token: " + token)
+	req.Header.Set("Authorization", "Token: "+token)
 	response = executeRequest(req)
 
 	checkResponseCode(t, http.StatusBadRequest, response.Code)
@@ -164,7 +184,7 @@ func TestScanRegistry(t *testing.T) {
 
 	payload := []byte(``)
 	req, _ := http.NewRequest("GET", "/registries/1/scan", bytes.NewBuffer(payload))
-	req.Header.Set("Authorization", "Token: " + token)
+	req.Header.Set("Authorization", "Token: "+token)
 	response := executeRequest(req)
 
 	checkResponseCode(t, 202, response.Code)
@@ -181,97 +201,29 @@ func TestScanRegistry(t *testing.T) {
 		Registry: "docker.io", Org: "johnptobe", JobName: "my-job", Title: "My first job",
 		Maintainer: "John Doe", Email: "jdoe@example.com", MaintOrg: "E-corp",
 		Description: "Reads an HDF5 file and outputs two TIFF images, a CSV and manifest containing cell_count",
-		JobVersion: "0.1.0", PackageVersion: "0.1.0"}
+		JobVersion:  "0.1.0", PackageVersion: "0.1.0"}
 	if fmt.Sprint(m[0]) != fmt.Sprint(testImage) {
 		t.Errorf("Expected image to be %v. Got '%v'", testImage, m[0])
 	}
 
 	req, _ = http.NewRequest("GET", "/registries/test/scan", bytes.NewBuffer(payload))
-	req.Header.Set("Authorization", "Token: " + token)
+	req.Header.Set("Authorization", "Token: "+token)
 	response = executeRequest(req)
 
 	checkResponseCode(t, http.StatusBadRequest, response.Code)
 }
 
-func TestSearchImages(t *testing.T) {
-	clearTable()
-
-	addRegistry()
-
-	payload := []byte(``)
-	req, _ := http.NewRequest("GET", "/registries/1/scan", bytes.NewBuffer(payload))
-	req.Header.Set("Authorization", "Token: " + token)
-	response := executeRequest(req)
-
-	checkResponseCode(t, 202, response.Code)
-
-	req, _ = http.NewRequest("GET", "/images/search/my-job-0.1.0", bytes.NewBuffer(payload))
-	response = executeRequest(req)
-
-	checkResponseCode(t, http.StatusOK, response.Code)
-
-	m := []models.SimpleImage{}
-	json.Unmarshal(response.Body.Bytes(), &m)
-
-	testImage := models.SimpleImage{ID: 1, RegistryId: 1, Name: "my-job-0.1.0-seed:0.1.0",
-		Registry: "docker.io", Org: "johnptobe", JobName: "my-job", Title: "My first job",
-		Maintainer: "John Doe", Email: "jdoe@example.com", MaintOrg: "E-corp",
-		Description: "Reads an HDF5 file and outputs two TIFF images, a CSV and manifest containing cell_count",
-		JobVersion: "0.1.0", PackageVersion: "0.1.0"}
-	if fmt.Sprint(m[0]) != fmt.Sprint(testImage) {
-		t.Errorf("Expected image to be %v. Got '%v'", testImage, m[0])
-	}
-
-	req, _ = http.NewRequest("GET", "/images/search/asdfasdf", bytes.NewBuffer(payload))
-	response = executeRequest(req)
-
-	checkResponseCode(t, http.StatusOK, response.Code)
-	json.Unmarshal(response.Body.Bytes(), &m)
-
-	if len(m) != 0 {
-		t.Errorf("Expected emtpy image list. Got %d results.", len(m))
-	}
-}
-
-func TestImageManifest(t *testing.T) {
-	clearTable()
-
-	addRegistry()
-
-	payload := []byte(``)
-	req, _ := http.NewRequest("GET", "/registries/1/scan", bytes.NewBuffer(payload))
-	req.Header.Set("Authorization", "Token: " + token)
-	response := executeRequest(req)
-
-	checkResponseCode(t, 202, response.Code)
-
-	req, _ = http.NewRequest("GET", "/images/1/manifest", bytes.NewBuffer(payload))
-	response = executeRequest(req)
-
-	checkResponseCode(t, http.StatusOK, response.Code)
-
-	m := objects.Seed{}
-	json.Unmarshal(response.Body.Bytes(), &m)
-
-	testManifest := objects.SeedFromManifestFile("seed.manifest.json")
-
-	mStr := fmt.Sprintf("%v", m)
-	testStr := fmt.Sprintf("%v", testManifest)
-	if mStr != testStr {
-		t.Errorf("Expected manifest to be %v. Got '%v'", testManifest, m)
-	}
-}
-
 func clearTable() {
-	db := GetDb()
 	db.Exec("DELETE FROM RegistryInfo")
 	db.Exec("DELETE FROM Image")
 	db.Exec("DELETE FROM sqlite_sequence")
+	db.Exec("DELETE FROM Job")
+	db.Exec("DELETE FROM JobVersion")
 }
 
 func executeRequest(req *http.Request) *httptest.ResponseRecorder {
 	rr := httptest.NewRecorder()
-	GetRouter().ServeHTTP(rr, req)
+	router.ServeHTTP(rr, req)
 
 	return rr
 }
@@ -286,7 +238,7 @@ func addRegistry() {
 	payload := []byte(`{"name":"dockerhub", "url":"https://hub.docker.com", "org":"johnptobe", "username":"", "password": ""}`)
 
 	req, _ := http.NewRequest("POST", "/registries/add", bytes.NewBuffer(payload))
-	req.Header.Set("Authorization", "Token: " + token)
+	req.Header.Set("Authorization", "Token: "+token)
 	executeRequest(req)
 }
